% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{add_errors}
\alias{add_errors}
\title{Add a given proportion of sequencing/damage errors to a set of samples.
samples <- c("arch0", "arch1")}
\usage{
add_errors(gt, rate, samples)
}
\description{
Add a given proportion of sequencing/damage errors to a set of samples.
samples <- c("arch0", "arch1")
}
