% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tmrca.R
\name{add_step1_ci}
\alias{add_step1_ci}
\title{Take A00 TMRCA and mutation rate estimates and simulate
confidence intervals by bootstraping for both of them.}
\usage{
add_step1_ci(df, per_ind)
}
\description{
Take A00 TMRCA and mutation rate estimates and simulate
confidence intervals by bootstraping for both of them.
}
