% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tmrca.R
\name{add_step2_ci}
\alias{add_step2_ci}
\title{Take archaic Y chromosome TMRCA estimates and simulate
confidence intervals by bootstraping of branch counts.}
\usage{
add_step2_ci(df, per_ind)
}
\description{
Take archaic Y chromosome TMRCA estimates and simulate
confidence intervals by bootstraping of branch counts.
}
