% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tmrca.R
\name{calculate_ci}
\alias{calculate_ci}
\title{Calculate empirical confidence intervals from bootstrapped
values of simulated branch counts.}
\usage{
calculate_ci(df, per_ind)
}
\description{
Calculate empirical confidence intervals from bootstrapped
values of simulated branch counts.
}
