% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.R
\name{calculate_fitness}
\alias{calculate_fitness}
\title{Calculate genetic fitness for each simulated Y chromosome.}
\usage{
calculate_fitness(f, pop = "")
}
\description{
Calculate genetic fitness for each simulated Y chromosome.
}
