% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmrca.R
\name{calculate_tAfrRef}
\alias{calculate_tAfrRef}
\title{Calculate T_MRCA of all pairs of Africans and non-Africans, using a set
of EMHs of known age for branch shortening-based mutation rate estimation.}
\usage{
calculate_tAfrRef(gt, samples)
}
\description{
Calculate T_MRCA of all pairs of Africans and non-Africans, using a set
of EMHs of known age for branch shortening-based mutation rate estimation.
}
