% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmrca.R
\name{calculate_tarch}
\alias{calculate_tarch}
\title{Calculate T_MRCA of each of the specified archaic individuals
with modern human Y chromosomes.}
\usage{
calculate_tarch(gt, samples, tafr)
}
\description{
Calculate T_MRCA of each of the specified archaic individuals
with modern human Y chromosomes.
}
