% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmrca.R
\name{calculate_tArchRef}
\alias{calculate_tArchRef}
\title{Calculate T_MRCA of each of the specified archaic individuals
with modern human Y chromosomes.}
\usage{
calculate_tArchRef(gt, samples, tAfrRef)
}
\description{
Calculate T_MRCA of each of the specified archaic individuals
with modern human Y chromosomes.
}
