% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_coverage}
\alias{get_coverage}
\title{Read coverage info for a given BAM file in a defined set of regions.}
\usage{
get_coverage(bed, bam)
}
\description{
Read coverage info for a given BAM file in a defined set of regions.
}
