% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{read_genotypes}
\alias{read_genotypes}
\title{Read genotypes from a VCF file, returning a data frame object.}
\usage{
read_genotypes(archaic, capture, mindp, maxdp = 0.98, var_only = FALSE,
  tv_only = FALSE, bed_filter = NA)
}
\arguments{
\item{archaic}{Name of an archaic human sample, or a path to its VCF.}

\item{capture}{Capture set (full, lippold, exome).}

\item{mindp}{Minimum coverage at each site.}

\item{maxdp}{Maximum coverage at each site (specified as a proportion of an
upper tail of the entire coverage distribution).}
}
\description{
Read genotypes from a VCF file, returning a data frame object.
}
