% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_info}
\alias{read_info}
\title{Generate table of sample names and their population assignments and ages.}
\usage{
read_info(gt)
}
\description{
Generate table of sample names and their population assignments and ages.
}
