% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_siminfo}
\alias{read_siminfo}
\title{Generate table of sample names and their population assignments and ages
from the simulated data.}
\usage{
read_siminfo(gt)
}
\description{
Generate table of sample names and their population assignments and ages
from the simulated data.
}
