% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmrca.R
\name{site_patterns}
\alias{site_patterns}
\title{Assign mutations to all possible branches of a tree with 4 leaves.}
\usage{
site_patterns(df, w, x, y, z)
}
\description{
Assign mutations to all possible branches of a tree with 4 leaves.
}
