% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snp_props}
\alias{snp_props}
\title{Calculate proportion of different classes of SNPs observed in each sample.}
\usage{
snp_props(gt, normalize = NA, remove = NA)
}
\arguments{
\item{normalize}{Normalize counts by some SNP class? Either FALSE or "T-C", etc.}
}
\description{
Calculate proportion of different classes of SNPs observed in each sample.
}
