% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{read_genotypes}
\alias{read_genotypes}
\title{Read genotypes from a VCF file, returning a data frame object.}
\usage{
read_genotypes(archaic, highcov, mindp, maxdp = 0.975,
  var_only = FALSE, tv_only = FALSE)
}
\arguments{
\item{archaic}{Path to a low-coverage archaic Y chromosome VCF.}

\item{highcov}{Path to a merged high-coverage Y chromosome VCF.}

\item{mindp}{Minimum coverage at each site.}

\item{maxdp}{Maximum coverage at each site (specified as a proportion of an
upper tail of the entire coverage distribution).}
}
\description{
Read genotypes from a VCF file, returning a data frame object.
}
