% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{read_gt}
\alias{read_gt}
\title{Read genotypes from a VCF file, returning a data frame object.}
\usage{
read_gt(path, mindp = 0, var_only = FALSE, tv_only = FALSE,
  exclude = NA)
}
\description{
Read genotypes from a VCF file, returning a data frame object.
}
